refactor: change arquiteture and add motion
