+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "axios": "^1.3.4",
     "chart.js": "^4.2.1",
+    "framer-motion": "^10.10.0",
     "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
