--- src/components/body/cardRepos/cardsRepos.sass
@@ -1,45 +1,0 @@
-@import '../../style/variables/colors'
-@import '../../style/variables/fonts'
-
-.titleReposGit
-    font-family: $Primary-Font
-    color: $Primary_Detail
-    font-size: 2.5rem
-    user-select: none
-
-.cardContainer
-    $White: #ffffff1a
-    $Transparent: #ffffff00
-    background: linear-gradient(90deg,  $White -10%, $Transparent 5%, $Transparent 95%, $White 110%)
-    border-radius: 35px
-    text-align: center
-    width: 100%
-    height: 100%
-    padding: 0px 15px
-    display: flex
-    align-items: center
-    justify-content: space-between
-
-.arrow
-
-    color: $Background
-    font-size: 4rem
-    user-select: none
-    transition: 0.2s
-    &:hover
-        text-shadow: 0px 0px 8px $Background
-        cursor: pointer
-
-.cardsBox
-    height: 100%
-    width: 100%
-    justify-content: space-between
-    display: flex
-    overflow-x: auto
-    scroll-behavior: smooth
-    user-select: none
-    // justify-content: center
-    // arrow display: none
-
-.cardsBox::-webkit-scrollbar 
-    display: none
