+++ src/components/buttons/ButtonCards.sass
@@ -3,7 +3,7 @@
 
 .buttonRepos 
   font-family: $Primary-Font
-  font-size: 15px
+  font-size: 16px
   padding: 0.7em 2.7em
   letter-spacing: 0.06em
   position: relative
@@ -15,6 +15,7 @@
   background: linear-gradient(to right, rgba(27, 253, 156, 0.1) 1%, transparent 40%,transparent 60% , rgba(27, 253, 156, 0.1) 100%)
   color: $Primary_dark
   box-shadow: inset 0 0 10px rgba(27, 253, 156, 0.4), 0 0 9px 3px rgba(27, 253, 156, 0.1)
+  cursor: pointer
 
 .buttonRepos:hover 
   color: $Primary_dark
