+++ src/components/cards/CardsDark.jsx
@@ -1,12 +1,13 @@
 import ButtonCards from "../buttons/ButtonCards"
 import { Fragment } from "react"
+import { motion } from "framer-motion"
 import "./CardsDark.sass"
 
 function CardsDark(props) {
     return Array.from({ length: props.length }, (v, k) => k)
         .map((i) => ([
         <Fragment key={props.id[i][0]}>
-            <div className="cardDarkBorder" >
+            <motion.div className="cardDarkBorder" whileTap={{cursor: "grabbing"}} >
                 <div className="cardDarkBody">
                     <h1 className="cardDarkTitle">{props.title[i][1]}</h1>
                     <p className="cardDarkParagraph">{props.text[i][3] ?? "Readme"}</p>
@@ -15,7 +16,7 @@ function CardsDark(props) {
                         link={props.link[i][2]}
                     />
                 </div>
-            </div>
+            </motion.div>
         </Fragment> ]
         ))
 }
