+++ src/components/cards/CardsDark.sass
@@ -3,13 +3,14 @@
 
 .cardDarkBorder
     min-width: 250px
-    width: 25%
+    width: 20%
     height: 300px
     background-image: linear-gradient(163deg, $Primary_dark 0%, $Secundary_dark 100%)
     border-radius: 20px
     margin: 50px
     transition: all .3s
     flex: none
+    cursor: grab
 
 .cardDarkBody
     width: 100%
