+++ src/feature/body/body.sass
@@ -1,4 +1,4 @@
-@import '../style/variables/colors'
+@import '../../components/style/variables/colors'
 
 .body
     max-width: 100%
