+++ src/feature/body/graphics/Graphics.jsx
@@ -2,7 +2,7 @@ import "./graphics.sass"
 import requestLang from "../../../service/setGithubLangRepos"
 import PieChart from "./Piechart"
 import React, { useEffect, useState, Fragment } from 'react'
-import colors from "../../style/variables/colors"
+import colors from "../../../components/style/variables/colors"
 
 function Graphics(){
