+++ src/feature/body/graphics/Piechart.jsx
@@ -1,6 +1,6 @@
 import { Pie } from "react-chartjs-2";
 import { Chart } from 'chart.js/auto'
-import  colors from "../../style/variables/colors";
+import  colors from "../../../components/style/variables/colors";
 
 const { colorThema } = colors
