+++ src/feature/body/summary/summary.sass
@@ -1,6 +1,6 @@
-@import "../../style/variables/colors"
-@import '../../style/variables/fonts'
-@import '../../style/variables/keyframes'
+@import "../../../components/style/variables/colors"
+@import '../../../components/style/variables/fonts'
+@import '../../../components/style/variables/keyframes'
 
 .summary
     width: 80%
