+++ src/feature/footer/footer.sass
@@ -1,5 +1,5 @@
-@import '../style/variables/colors'
-@import '../style/variables/fonts'
+@import '../../components/style/variables/colors'
+@import '../../components/style/variables/fonts'
 
 .footerBox
     width: 100%
